feat: release 1.1 documentation
