+++ crypto_signal/crypto_signal/settings.py
@@ -38,6 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.messages',
     'django.contrib.staticfiles',
     'crypto_track',
+    'django_extensions',
 ]
 
 MIDDLEWARE = [
